pyproject.toml
[project]
name = "telephony-agent-004"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.0.17",
    "livekit-plugins-deepgram>=1.0.17",
    "livekit-plugins-noise-cancellation>=0.2.1",
    "livekit-plugins-openai>=0.10.9",
    "livekit-plugins-silero>=1.0.17",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
